add function comp to page level
